# docs/openapi.yaml
openapi: 3.1.0
info:
  title: DFSP API
  description: |
    Decentralized File Sharing Platform API.
  version: 1.0.0
  contact:
    name: DFSP Team
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://api.dfsp.app
    description: Production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    TelegramChatId:
      type: apiKey
      in: header
      name: X-TG-Chat-Id
      description: Telegram chat ID for bot endpoints
  schemas:
...
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: true
...
    PublicLinkCreateIn:
      type: object
      properties:
        version:
          type: integer
          nullable: true
        name_override:
          type: string
          nullable: true
        mime_override:
          type: string
          nullable: true
        ttl_sec:
          type: integer
          nullable: true
        max_downloads:
          type: integer
          nullable: true
        pow:
          type: object
          properties:
            enabled:
              type: boolean
            difficulty:
              type: integer
    PublicLinkCreateOut:
      type: object
      properties:
        token:
          type: string
        expires_at:
          type: string
          format: date-time
          nullable: true
        policy:
          $ref: '#/components/schemas/PublicLinkPolicy'
    PublicLinkPolicy:
      type: object
      properties:
        max_downloads:
          type: integer
          nullable: true
        pow_difficulty:
          type: integer
          nullable: true
        one_time:
          type: boolean
    PublicMetaOut:
      type: object
      properties:
        name:
          type: string
        size:
          type: integer
        mime:
          type: string
          nullable: true
        cid:
          type: string
          nullable: true
        fileId:
          type: string
        version:
          type: integer
          nullable: true
        expires_at:
          type: string
          format: date-time
          nullable: true
        policy:
          $ref: '#/components/schemas/PublicLinkPolicy'
    PowIn:
      type: object
      required: [nonce, solution]
      properties:
        nonce:
          type: string
        solution:
          type: string

    # Bot schemas
    BotLinkItem:
      type: object
      properties:
        address:
          type: string
        is_active:
          type: boolean
    BotLinksResponse:
      type: object
      properties:
        links:
          type: array
          items:
            $ref: '#/components/schemas/BotLinkItem'
...
paths:
...
  /files/{file_id_hex}/public-links:
    post:
      tags: [public_links]
      summary: Create public link for file
      operationId: createPublicLink
      security:
        - BearerAuth: []
      parameters:
        - name: file_id_hex
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublicLinkCreateIn'
      responses:
        '201':
          description: Public link created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicLinkCreateOut'

    get:
      tags: [public_links]
      summary: List public links for file
      operationId: listPublicLinks
      security:
        - BearerAuth: []
      parameters:
        - name: file_id_hex
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of public links
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object

  # Grants
...
  /tg/link:
    delete:
      tags: [Telegram]
      summary: Unlink Telegram account
      operationId: tgUnlink
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Unlinked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
...
  /bot/links:
    get:
      tags: [Bot]
      summary: List linked wallets
      operationId: botListLinks
      security:
        - TelegramChatId: []
      responses:
        '200':
          description: List of links
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BotLinksResponse'

    post:
      tags: [Bot]
      summary: Create wallet link
      operationId: botCreateLink
      security:
        - TelegramChatId: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [address]
              properties:
                address:
                  type: string
                make_active:
                  type: boolean
      responses:
        '200':
          description: Link created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BotLinksResponse'
...
  /bot/links/{address}:
    delete:
      tags: [Bot]
      summary: Remove wallet link
      operationId: botRemoveLink
      security:
        - TelegramChatId: []
      parameters:
        - name: address
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Link removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BotLinksResponse'
...
  /bot/profile:
    get:
      tags: [Bot]
      summary: Get user profile
      operationId: botProfile
      security:
        - TelegramChatId: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  address:
                    type: string
                  display_name:
                    type: string
                    nullable: true
                  files_count:
                    type: integer
                  grants_received:
                    type: integer
                  grants_given:
                    type: integer
...
  /public/{token}/meta:
    get:
      tags: [public_links]
      summary: Get public link metadata
      operationId: publicMeta
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Public link metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicMetaOut'
        '404':
          description: Not found
        '410':
          description: Expired or revoked
  /public/{token}/pow:
    post:
      tags: [public_links]
      summary: Submit PoW solution for public link
      operationId: publicPow
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PowIn'
      responses:
        '200':
          description: PoW accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
        '400':
          description: Bad solution
  /public/{token}/content:
    get:
      tags: [public_links]
      summary: Download public link content
      operationId: publicContent
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: File content
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '403':
          description: PoW required
        '410':
          description: Expired or revoked